+++ lib/audioworklet-registry.ts
@@ -31,17 +31,3 @@ export const registeredWorklets: Map<
   AudioContext,
   Record<string, WorkletGraph>
 > = new Map();
-
-export const createWorketFromSrc = (
-  workletName: string,
-  workletSrc: string
-) => {
-  const script = new Blob(
-    [`registerProcessor("${workletName}", ${workletSrc})`],
-    {
-      type: 'application/javascript',
-    }
-  );
-
-  return URL.createObjectURL(script);
-};