+++ lib/worklets/audio-processing.ts
@@ -1,24 +1,9 @@
+/// <reference types="audioworklet" />
+
 /**
  * @license
  * SPDX-License-Identifier: Apache-2.0
 */
-/**
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-const AudioRecordingWorklet = `
 class AudioProcessingWorklet extends AudioWorkletProcessor {
 
   // send and clear buffer every 2048 samples, 
@@ -27,6 +12,7 @@ class AudioProcessingWorklet extends AudioWorkletProcessor {
 
   // current write index
   bufferWriteIndex = 0;
+  hasAudio: boolean;
 
   constructor() {
     super();
@@ -35,9 +21,9 @@ class AudioProcessingWorklet extends AudioWorkletProcessor {
 
   /**
    * @param inputs Float32Array[][] [input#][channel#][sample#] so to access first inputs 1st channel inputs[0][0]
-   * @param outputs Float32Array[][]
+   * @param outputs Float32Array[][][]
    */
-  process(inputs) {
+  process(inputs: Float32Array[][]) {
     if (inputs[0].length) {
       const channel0 = inputs[0][0];
       this.processChunk(channel0);
@@ -55,7 +41,7 @@ class AudioProcessingWorklet extends AudioWorkletProcessor {
     this.bufferWriteIndex = 0;
   }
 
-  processChunk(float32Array) {
+  processChunk(float32Array: Float32Array) {
     const l = float32Array.length;
     
     for (let i = 0; i < l; i++) {
@@ -72,6 +58,5 @@ class AudioProcessingWorklet extends AudioWorkletProcessor {
     }
   }
 }
-`;
 
-export default AudioRecordingWorklet;
+registerProcessor('audio-processing-worklet', AudioProcessingWorklet);